+++ inst/scripts/examples/all_funds.R
@@ -14,7 +14,9 @@ spec_src("dm_spec.R")
 spec_src("em_spec.R")
 spec_src("usa_sel_spec.R")
 spec_src("exus_spec.R")
+spec_src("emu_spec.R")
 spec_src("scv_spec.R")
+spec_src("estx50_spec.R")
 
 xlm_dir <- file.path("data", "xlm")